osa 4 done

+++ osa4/controllers/blogs.js
@@ -3,14 +3,6 @@ const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-        return authorization.substring(7)
-    }
-    return null
-}
-
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs.map(blog => blog.toJSON()))
@@ -19,7 +11,7 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response, next) => {
     try {
         const body = request.body
-        const token = getTokenFrom(request)
+        const token = request.token
 
         const decodedToken = jwt.verify(token, process.env.SECRET)
         if (!token || !decodedToken.id) {
@@ -40,13 +32,25 @@ blogsRouter.post('/', async (request, response, next) => {
         await user.save()
         response.status(201).json(savedBlog.toJSON())
     } catch (exception) {
-        console.log(exception)
         next(exception)
     }
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
     try {
+        const token = request.token
+
+        const decodedToken = jwt.verify(token, process.env.SECRET)
+        if (!token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
+        const blog = await Blog.findById(request.params.id)
+        console.log(blog.user, decodedToken.id)
+        if ( blog.user.toString() !== decodedToken.id.toString() ) {
+            return response.status(401).json({ error: 'only blog author can remove blog' })
+        }
+
         await Blog.findByIdAndRemove(request.params.id)
         response.status(204).end()
     } catch (exception) {
